Rename api.js to register_participant.js
